// apps/web/src/lib/api/__generated__/schemas/vocabularyCategory.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * WeeklyRun API
 * WeeklyRun - AI Meeting Orchestration API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Category of vocabulary term.
 */
export type VocabularyCategory = typeof VocabularyCategory[keyof typeof VocabularyCategory];


export const VocabularyCategory = {
  terminology: 'terminology',
  abbreviation: 'abbreviation',
  name: 'name',
  other: 'other',
} as const;